# openapi.yaml
openapi: 3.0.0
info:
  title: CivicConnect API
  description: API specification generated from implemented Go backend routes.
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/issues:
    get:
      summary: List public issues
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radiusMeters
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Issue list
...
  /api/v1/issues/{id}:
    get:
      summary: Public issue by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Issue detail
        '404':
          description: Not found
  /api/v1/citizen/issues:
    get:
      summary: Citizen nearby issues
      security:
        - bearerAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radiusMeters
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Citizen issue list
...
  /api/v1/citizen/issues/{id}:
    get:
      summary: Citizen issue by id
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Citizen issue detail
    delete:
      summary: Delete citizen issue
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Issue deleted
        '404':
          description: Issue not found
  /api/v1/citizen/issues/{id}/support:
    post:
      summary: Support issue
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Support updated
        '409':
          description: Duplicate support
  /api/v1/citizen/issues/{id}/confirm-resolution:
    post:
      summary: Confirm issue resolution
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confirmed
  /api/v1/citizen/issues/{id}/flags:
    post:
      summary: Create issue flag
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        '201':
          description: Flag created
        '409':
          description: Duplicate flag
  /api/v1/citizen/issues/{id}/flag:
    post:
      summary: Create issue flag (alias)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '201':
          description: Flag created
        '409':
          description: Duplicate flag
...
  /api/v1/authority/issues:
    get:
      summary: List authority issues
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Authority issues
  /api/v1/authority/issues/{id}:
    get:
      summary: Get authority issue detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Authority issue detail
...